--- app/business/services/invoice/revenue.command.ts
@@ -1,4 +1,0 @@
-// server action을 사용해 데이터를 mutation 하는 부분
-// https://nextjs.org/docs/app/building-your-application/data-fetching/server-actions-and-mutations
-// sample code는 못작성했음 revenue data에 관한 mutation 로직은 여기서 작성
-'use server';
