+++ app/mocks/handlers/result-handler.mock.ts
@@ -2,7 +2,7 @@ import { HttpResponse, http, delay } from 'msw';
 import { API_PATH } from '../../business/api-path';
 import { mockDatabase } from '../db.mock';
 import { ErrorResponseData } from '@/app/utils/http/http-error-handler';
-import { CreditResponse, ResultCategoryDetailResponse } from '@/app/business/services/result/result.type';
+import { CreditResponse, ResultCategoryDetailResponse } from '@/app/store/querys/result';
 
 export const resultHandlers = [
   http.get<never, never, ResultCategoryDetailResponse | ErrorResponseData>(
